web_app_gcp/app/v4/functions/thumbnail-worker/requirements.txt
functions-framework==3.8.0
google-cloud-storage==2.17.0
Pillow==10.4.0
